Updating WCS bbox and fixing terrainWidth
WCS bbox needs to cover the whole "Gløshaugen".
(If using stored texture mode instead of WMS, the WCS bbox also needs to
correspond with the texture bbox. This is the case after this update)
In addition the terrainWidth as based on the verticesRatio, when it
should have been length ratio. This did not affect non WCS mode, since
the resolution there is equal in north/south and west/east